shuffle accessors that return a geography to their own file
