+++ src/Makevars.in
@@ -106,6 +106,7 @@ OBJECTS = $(ABSL_LIBS) \
      s2-geography/capi/tessellator.o \
      s2-geography/capi/status.o \
      s2-geography/accessors.o \
+     s2-geography/accessors-geog.o \
      s2-geography/linear-referencing.o \
      s2-geography/distance.o \
      s2-geography/build.o \
