+++ src/Makevars.win
@@ -93,6 +93,7 @@ S2LIBS = $(ABSL_LIBS) \
      s2-geography/linear-referencing.o \
      s2-geography/distance.o \
      s2-geography/accessors.o \
+     s2-geography/accessors-geog.o \
      s2-geography/build.o \
      s2-geography/coverings.o \
      s2-geography/geography.o \
