+++ src/s2-geography/accessors.hpp
@@ -2,7 +2,6 @@
 #pragma once
 
 #include "geography.hpp"
-#include "aggregator.hpp"
 
 namespace s2geography {
 
@@ -16,18 +15,5 @@ double s2_perimeter(const S2Geography& geog);
 double s2_x(const S2Geography& geog);
 double s2_y(const S2Geography& geog);
 bool s2_find_validation_error(const S2Geography& geog, S2Error* error);
-
-S2Point s2_centroid(const S2Geography& geog);
-std::unique_ptr<S2Geography> s2_boundary(const S2Geography& geog);
-
-class S2CentroidAggregator: public S2Aggregator<S2Point> {
-public:
-    void Add(const S2Geography& geog);
-    void Merge(const S2CentroidAggregator& other);
-    S2Point Finalize();
-
-private:
-    S2Point centroid_;
-};
 
 }
