+++ src/s2-geography/s2-geography.hpp
@@ -3,6 +3,7 @@
 
 #include "geography.hpp"
 #include "accessors.hpp"
+#include "accessors-geog.hpp"
 #include "linear-referencing.hpp"
 #include "distance.hpp"
 #include "build.hpp"
